+++ protobuf/default.nix
@@ -1,4 +1,4 @@
-{ base, protobuf, protoc-gen-doc, languages }:
+{ base, protobuf, protoc-gen-doc, languages, grpcurl, runCommand, writeScriptBin }:
 let
   allLanguages = builtins.attrValues languages;
 in
@@ -41,11 +41,35 @@ in
             })
           (builtins.filter (value: value ? fromProtobuf) languages)
         );
+        protoset = runCommand
+          "generate-protoset"
+          { }
+          ''
+            ${protobuf}/bin/protoc \
+              --proto_path=${src} \
+              --descriptor_set_out="$out" \
+              --include_imports \
+              ${src}/**/*.proto
+          '';
+
+      grpcurlWrapped = writeScriptBin "grpcurl" ''
+        case "$1" in
+          *help)
+            ${grpcurl}/bin/grpcurl --help
+            exit 0;;
+
+          list|describe)
+            ${grpcurl}/bin/grpcurl -protoset "${protoset}" "$@"
+            exit;;
+        esac
+        ${grpcurl}/bin/grpcurl -protoset "${protoset}" -plaintext "$@"
+      '';
     in
     base.mkComponent (langs // {
       inherit name;
       _default = src;
       protobuf = src;
+      tools = grpcurlWrapped;
       docs.api = base.mkDerivation {
         name = "${name}-generated-docs";
         inherit src;